refactor: user관련 api 파일구조 변경
