--- apps/client/src/pages/my/apis/use-user-info.ts
@@ -1,19 +1,0 @@
-import { useSuspenseQuery } from '@tanstack/react-query';
-import { instance } from '@shared/apis/api';
-import { END_POINT } from '@shared/constants/api';
-import { BaseResponse } from '@shared/types/api';
-
-import { USER_QUERY_OPTIONS } from './queries';
-import { UserProfile } from '../types/user-response';
-
-export const getUserProfile = async (): Promise<UserProfile> => {
-  const response = await instance.get<BaseResponse<UserProfile>>(
-    END_POINT.GET_USER_PROFILE,
-  );
-  return response.data.data;
-};
-
-export const useUserProfile = () => {
-  const { data } = useSuspenseQuery(USER_QUERY_OPTIONS.PROFILE());
-  return data;
-};
