+++ apps/client/src/pages/my/page/profile/my-profile.tsx
@@ -3,13 +3,13 @@ import { ARTISTS_DATA } from '@shared/mocks/artists-data';
 import { PERFORMANCE_DATA } from '@shared/mocks/performance-data';
 import { routePath } from '@shared/constants/path';
 
-import { useUserProfile } from '@pages/my/apis/use-user-info';
-import UserInfo from '../../components/profile/user-info';
-import Box from '../../components/profile/box';
-import NoArtistSection from '../../components/artist/no-artist-section';
-import NoConfetiSection from '../../components/confeti/no-confeti-section';
-import ArtistSection from '../../components/artist/artist-section';
-import ConfetiSection from '../../components/confeti/conteti-section';
+import { useUserProfile } from '@pages/my/hooks/use-user-info';
+import Box from '@pages/my/components/profile/box';
+import NoArtistSection from '@pages/my/components/artist/no-artist-section';
+import NoConfetiSection from '@pages/my/components/confeti/no-confeti-section';
+import ArtistSection from '@pages/my/components/artist/artist-section';
+import UserInfo from '@pages/my/components/profile/user-info';
+import ConfetiSection from '@pages/my/components/confeti/conteti-section';
 
 const MyProfile = () => {
   const profileData = useUserProfile();
