+++ apps/client/src/shared/types/user-response.ts
